Uploading v1.00 of the Mastery Clock Project
this a timer that will help you track total time spent on a project. Every time you start it, it will begin from where it last left off, it will also tell you on which day how long you spent and what task you spent it on. The idea behind naming it 'Mastery Clock' is because it goes to 10,000 hours - once it reaches that time it will break ;)